requirements.txt
azure-search-documents>=11.6.0b12
azure-core>=1.30.0
python-dotenv>=1.0.1
openai>=1.55.0
mcp>=1.13.0